requirements.txt
annotated-types==0.6.0
anyio==4.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
cobble==0.1.3
distro==1.8.0
docx2python==2.7.3
docx2txt==0.8
exceptiongroup==1.2.0
Flask==3.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
mammoth==1.6.0
MarkupSafe==2.1.3
openai==1.3.8
packaging==23.2
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
pytesseract==0.3.10
python-docx==1.1.0
python-dotenv==1.0.0
python-pptx==0.6.23
sniffio==1.3.0
soupsieve==2.5
tqdm==4.66.1
typing_extensions==4.8.0
Werkzeug==3.0.1
XlsxWriter==3.1.9
zipp==3.17.0
